added tests for day 16 answers

+++ src/aoc-2020-csharp/Day16/Day16.cs
@@ -6,5 +6,5 @@ public static class Day16
 
     public static int Part1() => 1;
 
-    public static int Part2() => 2;
+    public static long Part2() => 2;
 }
